move Set TechPreviewNoUpgrade on the target cluster to networking step
